feat(database): added get account history method
The dhive client was missing the ability to query the account history for a specific account. While you could access the call method from the client instance and make the JSON RPC call yourself "get_account_history" this puts Dhive closer on parity with the Steemjs client which this package is based off of.